backend/pyproject.toml
[project]
name = "kanban-studio"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20",
    "fastapi>=0.115",
    "httpx>=0.28",
    "itsdangerous>=2.2",
    "bcrypt>=4.0",
    "python-dotenv>=1.0",
    "sqlalchemy[asyncio]>=2.0",
    "uvicorn[standard]>=0.34",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "anyio>=4",
    "aiosqlite>=0.20",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
